Remove use of `Row#getAnyOption` from `FinaglePostgresDecoders`
This addresses a part of the cause of issue [#1844](https://github.com/getquill/quill/issues/1844) which is that the decoders that are not direct in `FinaglePostgresDecoders` rely on `Row#getAnyOption`. `Row#getAnyOption` relies on custom receiving functions being defined on the `PostgresClient`. postgres-finagle could define these by default (which it probably should), but by default it does not.  When custom receiving functions are not defined, `Row#getAnyOption` always returns `None`.

This change introduces an `orElse` method on `FinaglePostgresDecoder` which allows for composition of `ValueDecoder`s that was previous being accomplished by `PartialFunction`s. This solution removes the need for `ClassTag` (although we may want to keep it to keep error messages consistent) as well as the reliance on `Row#getAnyOption`.  It also removes the definition of `decoder[T]`.